refactored oauth login to work with production profile, added to readme
