--- src/main/java/org/fungover/system2024/security/SecurityConfig.java
@@ -1,54 +1,0 @@
-package org.fungover.system2024.security;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.MediaType;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
-import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
-import org.springframework.security.web.util.matcher.MediaTypeRequestMatcher;
-
-import static org.springframework.http.HttpMethod.GET;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    private final String baseUrl;
-
-    public SecurityConfig(@Value("${app.baseUrl}") String baseUrl) {
-        this.baseUrl = baseUrl;
-    }
-
-    @Bean
-    public SecurityFilterChain SecurityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .authorizeHttpRequests((authorizeRequests) -> authorizeRequests
-                        .requestMatchers(GET, "/").authenticated()
-                        .anyRequest().authenticated()
-                )
-                .oauth2Login(Customizer.withDefaults())
-                .logout((logout) -> logout
-                        .logoutUrl("/logout")
-                        .invalidateHttpSession(true)
-                        .clearAuthentication(true)
-                        .deleteCookies("JSESSIONID")
-                        .logoutSuccessHandler(logoutSuccessHandler())
-                )
-                .exceptionHandling((exceptions) -> exceptions
-                        .defaultAuthenticationEntryPointFor(
-                                new LoginUrlAuthenticationEntryPoint("/oauth2/authorization/github"),
-                                new MediaTypeRequestMatcher(MediaType.TEXT_HTML)
-                        ));
-        return http.build();
-    }
-
-    private LogoutSuccessHandler logoutSuccessHandler() {
-        return ((request, response, authentication) ->
-                response.sendRedirect(baseUrl + "/login"));
-    }
-}
