+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@ spring.sql.init.mode=always
 spring.datasource.url=jdbc:mysql://${DB_LOCAL}
 spring.datasource.username=${DB_USERNAME}
 spring.datasource.password=${DB_PASSWORD}
-spring.profiles.active=development
+spring.profiles.active=production
 
 spring.config.import=optional:file:./.env[.properties],optional:file:.env[.properties]
 spring.security.oauth2.client.registration.github.clientId=${GITHUB_CLIENT_ID}
